+++ examples/ex_benchmark_series_cpr.cpp
@@ -22,7 +22,7 @@ using namespace reservoir_simulator::linear_problem;
 namespace {
 
 template<typename SolverType>
-SolveResult solve_with_scalar(LinearProblem& lp, int maxIter, typename SolverType::params& prm)
+SolveResult solve_with_scalar(LinearProblem& lp, size_t maxIter, typename SolverType::params& prm)
 {
     prm.solver.maxiter = maxIter;
